+++ src/Common/Connection/Acceptor.cpp
@@ -20,7 +20,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #include "Acceptor.h"
 #include "PacketHandler.h"
 
-Acceptor::Acceptor(int port, AbstractPlayerFactory* apf) {
+Acceptor::Acceptor(short port, AbstractPlayerFactory* apf) {
 	abstractPlayerFactory = apf;
 
 	int iResult = WSAStartup(MAKEWORD(2,2), &wsaData);
