+++ src/WorldServer/Channels.h
@@ -27,7 +27,7 @@ class WorldServerAcceptPlayer;
 struct Channel {
 	WorldServerAcceptPlayer *player;
 	char ip[15];
-	int port;
+	short port;
 };
 
 class Channels {
@@ -37,7 +37,7 @@ public:
 			singleton = new Channels;
 		return singleton;
 	}
-	void registerChannel(WorldServerAcceptPlayer *player, int channel, char *ip, int port);
+	void registerChannel(WorldServerAcceptPlayer *player, int channel, char *ip, short port);
 	Channel * getChannel(int num);
 	int size();
 private:
